refactor: split mock connection into a package with vendor personality dispatch
